AutoLofiUploader-Backend/requirements.txt
# requirements.txt

flask
gspread
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client